+++ apps/ledger-live-desktop/src/renderer/screens/accounts/AccountRowItem/Placeholder.tsx
@@ -33,7 +33,7 @@ const Placeholder = () => {
   const { trackAddAccountEvent } = useAddAccountAnalytics();
 
   const { openAssetFlow } = useOpenAssetFlow(
-    ModularDrawerLocation.ADD_ACCOUNT,
+    { location: ModularDrawerLocation.ADD_ACCOUNT },
     MAD_SOURCE_PAGES.ACCOUNTS_PAGE,
   );
   const handleAddAccountClick = () => {